api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
asyncpg==0.30.0
pydantic==2.10.3
python-multipart==0.0.18
slowapi==0.1.9
confluent-kafka==2.6.1
# LangChain / LangGraph
langchain>=0.3.14
langchain-openai>=0.2.14
langgraph>=0.2.60
langchain-core>=0.3.29

# MCP Python SDK (client side)
mcp>=1.3.0

# Observability
arize-phoenix-otel>=0.7.0
openinference-instrumentation-langchain>=0.1.29
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp-proto-http>=1.29.0
prometheus-client>=0.21.0

# Utilities
tenacity>=9.0.0
structlog>=24.4.0

langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.1.18